requirements.txt
cryptography==46.0.3
Django==5.2.8
eth-brownie==1.21.0
eventlet==0.40.4
Flask-SocketIO==5.5.1
pipdeptree==2.13.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
setuptools==65.5.0
types-requests==2.32.4.20250913
websocket-client==1.9.0
